[Android] Remove isGridTabSwitcherEnabled and isTabGroupsAndroidEnabled
Chromium changes:
https://github.com/brave/chromium/commit/f300646db04f84970e8d67045ce05dda5b968a90

[Tab Switcher] Remove isGridTabSwitcherEnabled
Now that M117 is branched and TabGroupsContinuationAndroid is the
default everywhere the old flags for the initial development of the Grid
Tab Switcher can finally be cleaned up.

This is the fourth step which removes the isGridTabSwitcherEnabled
method everywhere. As a result there are also a number of cascading
side effects in distant files.

Bug: 1466158

https://github.com/brave/chromium/commit/9724e9d8e7a2c1b07cd20964de56ab95bbdda72f

[Tab Switcher] Remove isTabGroupsAndroidEnabled
Now that M117 is branched and TabGroupsContinuationAndroid is the
default everywhere the old flags for the initial development of the Grid
Tab Switcher can finally be cleaned up.

This is the fifth step which removes the isTabGroupsAndroidEnabled
method everywhere. As a result there are also a number of cascading
side effects outside the tab switcher proper.

Bug: 1466158